逛逛附近无数据

+++ api/src/main/java/com/toutiao/appV2/model/sellhouse/SellHouseSearchDomainResponse.java
@@ -15,7 +15,7 @@ public class SellHouseSearchDomainResponse {
     private List<SellHousesSearchDo> data;
 
     @ApiModelProperty(value = "总条数", name = "totalNum")
-    private Integer totalNum;
+    private long totalNum;
 
     @ApiModelProperty(value = "是否为猜你喜欢的数据", name = "isGuess")
     private Integer isGuess;
